requirements.txt
anyio==4.13.0
asgiref==3.9.1
attrs==26.1.0
CacheControl==0.14.4
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
colorama==0.4.6
cryptography==46.0.7
Django==5.2.6
django-cors-headers==4.9.0
django-cuser==2017.3.16
django-filter==25.1
django-jazzmin==3.0.4
django-tinymce==5.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
drf-spectacular==0.28.0
Faker==40.15.0
firebase_admin==7.4.0
google-api-core==2.30.3
google-auth==2.49.2
google-cloud-core==2.5.1
google-cloud-firestore==2.27.0
google-cloud-storage==3.10.1
google-crc32c==1.8.0
google-resumable-media==2.8.2
googleapis-common-protos==1.74.0
grpcio==1.80.0
grpcio-status==1.80.0
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
modeltranslation==0.25
msgpack==1.1.2
packaging==25.0
pillow==11.3.0
proto-plus==1.27.2
protobuf==6.33.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
PyJWT==2.10.1
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
qrcode==8.2
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
